% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_trial.R
\name{agd_trial}
\alias{agd_trial}
\title{Create an aggregate dataset from a trial}
\usage{
agd_trial(trial)
}
\arguments{
\item{trial}{An object of class 'tte'}
}
\description{
Create an aggregate dataset from a trial
}
