% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_trial.R
\name{aggregate.tte}
\alias{aggregate.tte}
\title{Create an aggregate dataset from a trial}
\usage{
\method{aggregate}{tte}(trial)
}
\arguments{
\item{trial}{An object of class 'tte'}
}
\description{
Create an aggregate dataset from a trial
}
