% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte.R
\name{median.tte}
\alias{median.tte}
\title{Median of a trial in a TTE object}
\usage{
\method{median}{tte}(trial, strat = "TRT")
}
\arguments{
\item{strat}{The strata to calculate medians for. Defaults to treatment}

\item{tial}{An object of class 'tte'}
}
\description{
Median of a trial in a TTE object
}
