% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medians.R
\name{median_difference}
\alias{median_difference}
\title{Differences in medians}
\usage{
median_difference(trial, strat = "TRT")
}
\arguments{
\item{trial}{An object of class 'tte'}

\item{strat}{The strata on which to calculate the differences between}
}
\description{
Differences in medians
}
