% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte.R
\name{plot.tte}
\alias{plot.tte}
\title{Plot a trial simulation}
\usage{
\method{plot}{tte}(
  trial,
  strat = "TRT",
  xlab = "Time",
  ylab = "Survival",
  type = "survival",
  legend = "Treatment",
  CI = TRUE,
  censor = TRUE,
  median = FALSE
)
}
\arguments{
\item{trial}{An object of class 'tte'}
}
\description{
Plot a trial simulation
}
