% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial.R
\name{simulate_trial}
\alias{simulate_trial}
\title{Simulate a trial}
\usage{
simulate_trial(
  n = 50,
  distribution = "weibull",
  covs = NULL,
  lambdas = 0.1,
  gammas = 1.5,
  pBeta = -0.5,
  sizebnom = 1L,
  pBinom = 0.5,
  maxt = NULL,
  trts = c("A", "B"),
  pMale = 0.51,
  tdeVal = 0.15,
  tdef = "log",
  seed = 1,
  ageMean = 65,
  ageSD = 2.5
)
}
\arguments{
\item{n}{Number of patients in total.}

\item{covs}{A data frame containing covariate information for each individual.}

\item{lambdas}{See [simsurv::simsurv].}

\item{gammas}{See [simsurv::simsurv].}

\item{maxt}{Maximim observation time.}

\item{trts}{Vector of treatments.}

\item{pbeta}{Log hazard ratio for treatment.}

\item{sizebinom}{For the [stats::rbinom function].}

\item{pbinom}{Probability of being assigned to a treatment.}
}
\value{
An object of class simtrial
}
\description{
Wrapper around [simsurv::simsurv] to simulate a clinical trial
}
