% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial.R
\name{simulate_trial}
\alias{simulate_trial}
\title{Simulate a trial}
\usage{
simulate_trial(
  n = 50,
  covs = NULL,
  lambdas = 0.1,
  gammas = 1.5,
  pbeta = -0.5,
  sizebnom = 1L,
  pbinom = 0.5,
  maxt = 10,
  trts = c("A", "B")
)
}
\arguments{
\item{n}{Number of patients in total.}

\item{covs}{A data frame containing covariate information for each individual.}

\item{lambdas}{See [simsurv::simsurv].}

\item{gammas}{See [simsurv::simsurv].}

\item{pbeta}{Log hazard ratio for treatment.}

\item{pbinom}{Probability of being assigned to a treatment.}

\item{maxt}{Maximim observation time.}

\item{trts}{Vector of treatments.}

\item{sizebinom}{For the [stats::rbinom function].}
}
\value{
An object of class simtrial
}
\description{
Wrapper around [simsurv::simsurv] to simulate a clinical trial
}
